Add files via upload
these committed changes fix issue#4 space traversal of k-fold. in this the obtained hyper parameter tuned model from PR  for #3 is used in KNN model and k-fold as well as its variant stratified k-fold is used for accuracy evaluation of the classification by KNN model by varying the no. of folds. the mean_score is used as evaluation metric.